+++ GothicModComposer.UI/Views/GmcSettings.xaml.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.IO;
+using System.Linq;
 using System.Windows;
 using System.Windows.Data;
 using GothicModComposer.UI.Models;
@@ -12,7 +14,7 @@ namespace GothicModComposer.UI.Views
     /// </summary>
     public partial class GmcSettings
     {
-		private const string WindowsStartupRegistryPath = "SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Run";
+        private const string WindowsStartupRegistryPath = "SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Run";
 
         public GmcSettings(GmcSettingsVM gmcSettingsVM)
         {
@@ -22,20 +24,24 @@ namespace GothicModComposer.UI.Views
 
             var collectionView = new ListCollectionView(gmcSettingsVM.GmcConfiguration.IniOverrides);
             collectionView.GroupDescriptions.Add(new PropertyGroupDescription(nameof(IniOverride.Section)));
-
+            gmcSettingsVM.IsLogDirectoryAvailable = Directory.Exists(gmcSettingsVM.LogsDirectoryPath) &&
+                                                    new DirectoryInfo(gmcSettingsVM.LogsDirectoryPath).GetFiles().Any();
             OverridesIniTable.ItemsSource = collectionView;
+            modBuildBtn.IsEnabled =
+                Directory.Exists(Path.Combine(gmcSettingsVM.GmcConfiguration?.Gothic2RootPath ?? string.Empty, ".gmc",
+                    "build"));
         }
 
         private void WindowsStartup_Checked(object sender, System.Windows.RoutedEventArgs e)
         {
-	        var exePath = System.Reflection.Assembly.GetExecutingAssembly().Location.Replace("dll", "exe");
+            var exePath = System.Reflection.Assembly.GetExecutingAssembly().Location.Replace("dll", "exe");
 
             try
             {
                 var registryKey = Registry.CurrentUser.OpenSubKey(WindowsStartupRegistryPath, true);
                 registryKey?.SetValue("GMC_2_UI", exePath);
             }
-            catch (Exception ex)
+            catch (Exception)
             {
                 MessageBox.Show("Error during adding application to Windows startup.");
             }
@@ -48,10 +54,20 @@ namespace GothicModComposer.UI.Views
                 var registryKey = Registry.CurrentUser.OpenSubKey(WindowsStartupRegistryPath, true);
                 registryKey?.DeleteValue("GMC_2_UI", false);
             }
-            catch (Exception ex)
+            catch (Exception)
             {
                 MessageBox.Show("Error during removing application from Windows startup");
             }
+        }
+
+        private void gothicRoot_MouseDoubleClick(object sender, System.Windows.Input.MouseButtonEventArgs e)
+        {
+            Clipboard.SetText(gothicRoot.Text);
+        }
+
+        private void gothicModRoot_MouseDoubleClick(object sender, System.Windows.Input.MouseButtonEventArgs e)
+        {
+            Clipboard.SetText(gothicModRoot.Text);
         }
     }
 }