+++ GothicModComposer/Commands/ExecuteGothicCommand.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics;
+using System.IO;
 using GothicModComposer.Models.Profiles;
 using GothicModComposer.Utils;
 using GothicModComposer.Utils.GothicSpyProcess;
@@ -26,7 +27,11 @@ namespace GothicModComposer.Commands
 
         private readonly IProfile _profile;
         private Process _gothicProcess;
-        private IndeterminateProgressBar _progressBar;
+
+        private IndeterminateProgressBar _rootProgressBar;
+        private ChildProgressBar _textureCompilationProgressBar;
+
+        private FileSystemWatcher _compiledTexturesFileWatcher;
 
         public ExecuteGothicCommand(IProfile profile, string killProcessMessage = null)
         {
@@ -47,24 +52,28 @@ namespace GothicModComposer.Commands
 
             Logger.Info($"Executing with kill process message: '{_killProcessMessage}'", true);
 
-            _gothicProcess = new Process
-            {
-                StartInfo = GetGothicProcessStartInfo()
-            };
+            _gothicProcess = GetGothicProcess();
 
             _gothicSpyProcessRunner.Run();
             _gothicSpyProcessRunner.Subscribe(Notify);
 
             Logger.Info($"{_gothicProcess.StartInfo.FileName} {_gothicProcess.StartInfo.Arguments}", true);
 
-            using (_progressBar = new IndeterminateProgressBar(
+            using (_rootProgressBar = new IndeterminateProgressBar(
                 $"Gothic2.exe process executed with arguments '{_gothicProcess.StartInfo.Arguments}'",
                 ProgressBarOptionsHelper.Get()))
             {
+                if (IsTextureCompilationRequired())
+                {
+                    StartRealTimeProgressOnTextureCompilation();
+                }
+
                 _gothicProcess.Start();
                 _gothicProcess.WaitForExit();
 
-                _progressBar.Finished();
+                _rootProgressBar.Finished();
+                _textureCompilationProgressBar?.Dispose();
+                _compiledTexturesFileWatcher?.Dispose();
             }
 
             _gothicSpyProcessRunner.Abort();
@@ -76,8 +85,8 @@ namespace GothicModComposer.Commands
         {
             Logger.zLog(message);
 
-            if (_progressBar != null)
-                _progressBar.Message = message;
+            if (_rootProgressBar != null)
+                _rootProgressBar.Message = message;
 
             if (_killProcessMessage is null || !message.Contains(_killProcessMessage))
                 return;
@@ -87,7 +96,7 @@ namespace GothicModComposer.Commands
             Logger.Info("Gothic process was terminated.", true);
         }
 
-        private ProcessStartInfo GetGothicProcessStartInfo()
+        private Process GetGothicProcess()
         {
             string arguments;
 
@@ -97,15 +106,38 @@ namespace GothicModComposer.Commands
             else
                 arguments = _profile.GothicArguments.ToString();
 
-
-            return new ProcessStartInfo
+            return new Process
             {
-                FileName = _profile.GothicFolder.GothicExeFilePath,
-                WorkingDirectory = _profile.GothicFolder.BasePath,
-                Arguments = arguments,
-                Verb = "runas", // Force to run the process as Administrator
-                UseShellExecute = false
+                StartInfo = new ProcessStartInfo
+                {
+                    FileName = _profile.GothicFolder.GothicExeFilePath,
+                    WorkingDirectory = _profile.GothicFolder.BasePath,
+                    Arguments = arguments,
+                    Verb = "runas", // Force to run the process as Administrator
+                    UseShellExecute = false
+                }
             };
+        }
+
+        private bool IsTextureCompilationRequired()
+            => _profile.GothicArguments.Contains(GothicArguments.ZReparseParameter) ||
+               _profile.GothicArguments.Contains(GothicArguments.ZTexConvertParameter);
+
+        private void StartRealTimeProgressOnTextureCompilation()
+        {
+            var numberOfTexturesToCompile = _profile.GothicFolder.GetNumberOfTexturesToCompile();
+            var counter = 1;
+
+            _textureCompilationProgressBar = _rootProgressBar?.Spawn(
+                numberOfTexturesToCompile, "Compiling textures", ProgressBarOptionsHelper.Get());
+
+            _compiledTexturesFileWatcher = new FileSystemWatcher(_profile.GothicFolder.CompiledTexturesPath);
+            _compiledTexturesFileWatcher.Created += (_, _) =>
+            {
+                _textureCompilationProgressBar?.Tick(
+                    $"Compiled {counter++} of {numberOfTexturesToCompile} textures");
+            };
+            _compiledTexturesFileWatcher.EnableRaisingEvents = true;
         }
     }
 }